[Optimize] Optimize WebsocketClient
	Use WebSocketTask to manage thread and socket.

issue: #23
